+++ src/enums/sessionStates.enum.ts
@@ -9,6 +9,7 @@ export enum SessionLogRecordType {
 export enum SessionStateType {
 	created = "created",
 	running = "running",
+	stopped = "stopped",
 	error = "error",
 	completed = "completed",
 }