+++ src/interfaces/projectView.interface.ts
@@ -13,6 +13,10 @@ export interface IProjectViewDelegate {
 	activateDeployment?: Callback<string>;
 	deactivateDeployment?: Callback<string>;
 	startSession?: Callback<StartSessionArgsType>;
+	deleteDeployment?: Callback<string>;
+	deleteSession?: Callback<string>;
+	displayErrorWithoutActionButton?: Callback<string>;
+	stopSession?: Callback<string>;
 }
 
 export interface IProjectView {