+++ src/utilities/vsCodeMessageHandler.utils.ts
@@ -1,12 +1,21 @@
 import { LoggerService } from "@services";
+import { WorkspaceConfig } from "@utilities/workspaceConfig.util";
 import { window } from "vscode";
 
 export class MessageHandler {
+	private static notificationsLevel = WorkspaceConfig.getFromWorkspace<string>("notificationsLevel", "");
+
 	static infoMessage(messageText: string): void {
+		if (MessageHandler.notificationsLevel !== "All") {
+			return;
+		}
 		window.showInformationMessage(messageText);
 	}
-	static errorMessage(messageText: string): void {
-		window.showErrorMessage(messageText, "View log").then((selection) => {
+	static errorMessage(messageText: string, displayShowLogButton = true): void {
+		if (MessageHandler.notificationsLevel === "None") {
+			return;
+		}
+		window.showErrorMessage(messageText, displayShowLogButton ? "View log" : "").then((selection) => {
 			if (selection === "View log") {
 				LoggerService.reveal();
 			}