+++ vscode-react/src/components/aKDeletePopper.component.tsx
@@ -1,0 +1,44 @@
+import React from "react";
+import { translate } from "@i18n";
+import { Player } from "@lottiefiles/react-lottie-player";
+import loaderAnimation from "@react-assets/media/catto-loader.json";
+import { AKButton } from "@react-components/aKButton.component";
+
+interface DeletePopperProps {
+	isDeletingInProcess: boolean;
+	onConfirm: () => void;
+	onDismiss: () => void;
+	translations: {
+		title: string;
+		subtitle: string;
+	};
+}
+
+export const DeletePopper: React.FC<DeletePopperProps> = ({
+	isDeletingInProcess,
+	onConfirm,
+	onDismiss,
+	translations,
+}) => (
+	<>
+		<div className="relative shadow-lg">
+			{isDeletingInProcess ? (
+				<Player src={loaderAnimation} className="player" loop autoplay />
+			) : (
+				<>
+					<div className="mb-3 text-left">
+						<strong>{translations.title}</strong>
+						<div>{translations.subtitle}</div>
+					</div>
+					<div className="flex">
+						<AKButton classes="bg-vscode-editor-background text-vscode-foreground" onClick={onDismiss}>
+							{translate().t("reactApp.general.no")}
+						</AKButton>
+						<div className="flex-grow" />
+						<AKButton onClick={onConfirm}>{translate().t("reactApp.general.yes")}</AKButton>
+					</div>
+				</>
+			)}
+		</div>
+	</>
+);