+++ vscode-react/src/components/aKDeploymentTableBody.component.tsx
@@ -2,35 +2,51 @@ import React, { useEffect, useRef, useState } from "react";
 import { DeploymentState, MessageType, SessionStateType } from "@enums";
 import { translate } from "@i18n";
 import { AKDeploymentState } from "@react-components";
-import { AKButton } from "@react-components/aKButton.component";
+import { DeletePopper, ExecutePopper, PopperComponent } from "@react-components";
 import { AKTableCell, AKTableRow } from "@react-components/AKTable";
+import { useAppState } from "@react-context/appState.context";
 import { useDeployments } from "@react-hooks";
+import { useIncomingMessageHandler } from "@react-hooks";
 import { getTimePassed, sendMessage } from "@react-utilities";
-import { cn } from "@react-utilities/cnClasses.utils";
 import { Deployment, SessionEntrypoint } from "@type/models";
-import { VSCodeDropdown } from "@vscode/webview-ui-toolkit/react";
-import { usePopper } from "react-popper";
 
 export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deployment[] }) => {
+	// State Hooks Section
 	const { selectedDeploymentId, entrypoints } = useDeployments();
-
-	useEffect(() => {
-		if (typeof selectedDeploymentId === "string") {
-			setSelectedDeployment(selectedDeploymentId);
-		}
-	}, [selectedDeploymentId]);
-
-	const referenceEl = useRef<HTMLDivElement | null>(null);
-	const popperEl = useRef<HTMLDivElement | null>(null);
-	const isDeploymentStateStartable = (deploymentState: number) =>
-		deploymentState === DeploymentState.INACTIVE_DEPLOYMENT || deploymentState === DeploymentState.DRAINING_DEPLOYMENT;
+	const [{ modalName }, dispatch] = useAppState();
+	const executePopperElementRef = useRef<HTMLDivElement | null>(null);
+	const deletePopperElementRef = useRef<HTMLDivElement | null>(null);
 	const [selectedFile, setSelectedFile] = useState<string>("");
 	const [selectedFunction, setSelectedFunction] = useState<string>("");
 	const [selectedEntrypoint, setSelectedEntrypoint] = useState<SessionEntrypoint>();
-
 	const [selectedDeployment, setSelectedDeployment] = useState("");
 	const [files, setFiles] = useState<Record<string, SessionEntrypoint[]>>();
 	const [functions, setFunctions] = useState<SessionEntrypoint[]>();
+	const [isDeletingInProcess, setIsDeletingInProgress] = useState(false);
+	const [deleteDeploymentId, setDeleteDeploymentId] = useState<string | null>(null);
+	const [displayedErrors, setDisplayedErrors] = useState<Record<string, boolean>>({});
+
+	// Local variable
+	const deleteDeploymentPopperTranslations = {
+		title: translate().t("reactApp.deployments.deletionApprovalQuestion"),
+		subtitle: translate().t("reactApp.deployments.deletionApprovalQuestionSubtitle"),
+	};
+
+	// Incoming Messages Handler
+	const handleDeploymentDeletedResponse = () => {
+		setIsDeletingInProgress(false);
+		hidePopper();
+	};
+
+	useIncomingMessageHandler({ handleDeploymentDeletedResponse });
+
+	// Functions Section
+	const showPopper = (popperId: string) => dispatch({ type: "SET_MODAL_NAME", payload: popperId });
+	const hidePopper = () => dispatch({ type: "SET_MODAL_NAME", payload: "" });
+
+	const isDeploymentStateStartable = (deploymentState: number) =>
+		deploymentState === DeploymentState.INACTIVE_DEPLOYMENT || deploymentState === DeploymentState.DRAINING_DEPLOYMENT;
+
 	const getSessionStateCount = (deployment: Deployment, state: string) => {
 		if (!deployment.sessionStats) {
 			return translate().t("reactApp.general.unknown");
@@ -38,22 +54,17 @@ export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deploymen
 		const session = deployment.sessionStats.find((s) => s.state === state);
 		return session ? session.count : 0;
 	};
-	const { attributes, styles } = usePopper(referenceEl.current, popperEl.current, {
-		placement: "bottom",
-		modifiers: [
-			{
-				name: "offset",
-				options: {
-					offset: [0, 10],
-				},
-			},
-		],
-	});
-	const [displayExecutePopper, setDisplayExecutePopper] = useState<boolean>(false);
-	const popperClasses = cn(
-		"flex-col z-30 bg-vscode-editor-background text-vscode-foreground",
-		"border border-gray-300 p-4 rounded-lg shadow-lg"
-	);
+
+	const handleFunctionChange = (event: string) => {
+		let entrypointForFunction;
+		try {
+			entrypointForFunction = JSON.parse(event);
+		} catch (error) {
+			console.error(error);
+		}
+		setSelectedEntrypoint(entrypointForFunction);
+		setSelectedFunction(event);
+	};
 
 	const deactivateBuild = (deploymentId: string) => {
 		sendMessage(MessageType.deactivateDeployment, deploymentId);
@@ -63,25 +74,12 @@ export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deploymen
 		sendMessage(MessageType.activateDeployment, deploymentId);
 	};
 
-	const togglePopper = () => {
-		setDisplayExecutePopper(true);
-	};
 	const getSessionsByDeploymentId = (deploymentId: string) => {
 		sendMessage(MessageType.selectDeployment, deploymentId);
 		setSelectedDeployment(deploymentId);
 	};
 
-	const [displayedErrors, setDisplayedErrors] = useState<Record<string, boolean>>({});
-
-	useEffect(() => {
-		if (entrypoints && Object.keys(entrypoints).length) {
-			setFiles(entrypoints);
-			setSelectedFile(Object.keys(entrypoints)[0]);
-			setFunctions(entrypoints[Object.keys(entrypoints)[0]]);
-			setSelectedFunction(JSON.stringify(entrypoints[Object.keys(entrypoints)[0]][0]));
-			setSelectedEntrypoint(entrypoints[Object.keys(entrypoints)[0]][0]);
-		}
-	}, [entrypoints]);
+	const isActive = (deploymentState: DeploymentState) => deploymentState === DeploymentState.ACTIVE_DEPLOYMENT;
 
 	const startSession = () => {
 		const lastDeployment = deployments![0];
@@ -106,19 +104,57 @@ export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deploymen
 
 		sendMessage(MessageType.startSession, startSessionArgs);
 
-		setDisplayExecutePopper(false);
+		hidePopper();
 	};
 
-	const handleFunctionChange = (event: string) => {
-		let entrypointForFunction;
-		try {
-			entrypointForFunction = JSON.parse(event);
-		} catch (error) {
-			console.error(error);
-		}
-		setSelectedEntrypoint(entrypointForFunction);
-		setSelectedFunction(event);
+	const deleteDeploymentConfirmed = () => {
+		sendMessage(MessageType.deleteDeployment, deleteDeploymentId);
+		setIsDeletingInProgress(true);
 	};
+
+	const deleteDeploymentDismissed = () => {
+		setIsDeletingInProgress(false);
+		setDeleteDeploymentId("");
+		hidePopper();
+	};
+
+	const showDeleteDeploymentPopper = (event: React.MouseEvent<HTMLDivElement>, deployment: Deployment) => {
+		if (deployment.state === DeploymentState.ACTIVE_DEPLOYMENT) {
+			sendMessage(MessageType.displayErrorWithoutActionButton, translate().t("reactApp.deployments.deleteDisabled"));
+			return;
+		}
+		const refElement = event.currentTarget;
+		showPopper("deploymentDelete");
+		deletePopperElementRef.current = refElement;
+		setDeleteDeploymentId(deployment.deploymentId);
+	};
+
+	// useEffects Section
+	useEffect(() => {
+		if (typeof selectedDeploymentId === "string") {
+			setSelectedDeployment(selectedDeploymentId);
+		}
+	}, [selectedDeploymentId]);
+
+	useEffect(() => {
+		hidePopper();
+	}, []);
+
+	useEffect(() => {
+		if (entrypoints && Object.keys(entrypoints).length) {
+			setFiles(entrypoints);
+			setSelectedFile(Object.keys(entrypoints)[0]);
+			setFunctions(entrypoints[Object.keys(entrypoints)[0]]);
+			setSelectedFunction(JSON.stringify(entrypoints[Object.keys(entrypoints)[0]][0]));
+			setSelectedEntrypoint(entrypoints[Object.keys(entrypoints)[0]][0]);
+		}
+	}, [entrypoints]);
+
+	useEffect(() => {
+		if (typeof selectedDeploymentId === "string") {
+			setSelectedDeployment(selectedDeploymentId);
+		}
+	}, [selectedDeploymentId]);
 
 	return (
 		deployments &&
@@ -131,6 +167,9 @@ export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deploymen
 					<div className="flex justify-center">
 						<AKDeploymentState deploymentState={deployment.state} />
 					</div>
+				</AKTableCell>
+				<AKTableCell onClick={() => getSessionsByDeploymentId(deployment.deploymentId)} classes={["cursor-pointer"]}>
+					{getSessionStateCount(deployment, SessionStateType.stopped)}
 				</AKTableCell>
 				<AKTableCell onClick={() => getSessionsByDeploymentId(deployment.deploymentId)} classes={["cursor-pointer"]}>
 					{getSessionStateCount(deployment, SessionStateType.running)}
@@ -148,9 +187,9 @@ export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deploymen
 					{deployment.deploymentId === deployments?.[0]?.deploymentId && (
 						<div
 							className="codicon codicon-redo mr-2 cursor-pointer"
-							ref={referenceEl}
-							title="Execute"
-							onClick={() => togglePopper()}
+							ref={executePopperElementRef}
+							title={translate().t("reactApp.deployments.execute")}
+							onClick={() => showPopper("deploymentExecute")}
 						></div>
 					)}
 					{isDeploymentStateStartable(deployment.state) ? (
@@ -164,60 +203,43 @@ export const AKDeploymentTableBody = ({ deployments }: { deployments?: Deploymen
 							onClick={() => deactivateBuild(deployment.deploymentId)}
 						></div>
 					)}
-
-					{displayExecutePopper && (
-						<div className="absolute w-screen h-screen" onClick={() => setDisplayExecutePopper(false)} />
-					)}
 					<div
-						ref={popperEl}
-						style={styles.popper}
-						{...attributes.popper}
-						className={cn(popperClasses, [{ invisible: !displayExecutePopper }])}
-					>
-						<div className="mb-3 text-left">
-							<strong className="mb-2">{translate().t("reactApp.deployments.executeFile")}</strong>
-							<VSCodeDropdown
-								value={selectedFile}
-								onChange={(e: any) => setSelectedFile(e.target.value)}
-								className="flex"
-							>
-								{files &&
-									Object.keys(files).map((file) => (
-										<option key={file} value={file}>
-											{file}
-										</option>
-									))}
-							</VSCodeDropdown>
-							{displayedErrors["triggerFile"] && <div className="text-red-500">Please choose trigger file</div>}
-						</div>
-						<div className="mb-3 text-left">
-							<strong className="mb-2">{translate().t("reactApp.deployments.executeEntrypoint")}</strong>
-							<VSCodeDropdown
-								value={selectedFunction}
-								onChange={(e: any) => handleFunctionChange(e.target.value)}
-								disabled={functions !== undefined && functions.length <= 1}
-								className="flex"
-							>
-								{functions &&
-									functions.map((func) => (
-										<option key={func.name} value={JSON.stringify(func)}>
-											{func.name}
-										</option>
-									))}
-							</VSCodeDropdown>
-							{displayedErrors["triggerFunction"] && <div className="text-red-500">Please choose trigger function</div>}
-						</div>
-						<div className="flex">
-							<AKButton
-								classes="bg-vscode-editor-background text-vscode-foreground"
-								onClick={() => setDisplayExecutePopper(false)}
-							>
-								{translate().t("reactApp.deployments.dismiss")}
-							</AKButton>
-							<div className="flex-grow" />
-							<AKButton onClick={() => startSession()}>{translate().t("reactApp.deployments.saveAndRun")}</AKButton>
-						</div>
-					</div>
+						className={`relative codicon codicon-trash ${
+							isActive(deployment.state) ? "cursor-not-allowed" : "cursor-pointer"
+						} ml-2 z-20`}
+						title={
+							isActive(deployment.state)
+								? translate().t("reactApp.deployments.deleteDisabled")
+								: translate().t("reactApp.deployments.delete")
+						}
+						onClick={(event) => showDeleteDeploymentPopper(event, deployment)}
+					></div>
+
+					{(modalName === "deploymentDelete" || modalName === "deploymentExecute") && (
+						<div className="absolute h-screen w-screen top-0 left-0 z-10" onClick={() => hidePopper()}></div>
+					)}
+
+					<PopperComponent visible={modalName === "deploymentDelete"} referenceRef={deletePopperElementRef}>
+						<DeletePopper
+							isDeletingInProcess={isDeletingInProcess}
+							onConfirm={() => deleteDeploymentConfirmed()}
+							onDismiss={() => deleteDeploymentDismissed()}
+							translations={deleteDeploymentPopperTranslations}
+						/>
+					</PopperComponent>
+					<PopperComponent visible={modalName === "deploymentExecute"} referenceRef={executePopperElementRef}>
+						<ExecutePopper
+							files={files!}
+							functions={functions!}
+							selectedFile={selectedFile}
+							selectedFunction={selectedFunction}
+							onFileChange={setSelectedFile}
+							onFunctionChange={handleFunctionChange}
+							onStartSession={startSession}
+							onClose={() => hidePopper()}
+							displayedErrors={displayedErrors}
+						/>
+					</PopperComponent>
 				</AKTableCell>
 			</AKTableRow>
 		))