+++ vscode-react/src/interfaces/incomingMessagesHandler.interface.ts
@@ -1,7 +1,16 @@
 import { Theme } from "@enums";
+import { DeploymentSectionViewModel, SessionSectionViewModel } from "@models";
+import { SessionEntrypoint } from "@type/models";
 
 export interface IIncomingMessagesHandler {
-	setThemeVisualType(themeKind: Theme | undefined): void;
-	setProjectName(projectName: string | undefined): void;
-	setResourcesDirState(projectFolder: boolean): void;
+	setThemeVisualType?(themeKind: Theme | undefined): void;
+	setProjectName?(projectName: string | undefined): void;
+	setResourcesDirState?(projectFolder: boolean): void;
+	handleSessionDeletedResponse?: (isDeleted: boolean) => void;
+	handleDeploymentDeletedResponse?: (isDeleted: boolean) => void;
+	setEntrypoints?: (value: Record<string, SessionEntrypoint[]> | undefined) => void;
+	setDeploymentsSection?: (value: DeploymentSectionViewModel | undefined) => void;
+	setSelectedDeploymentId?: (selectDeploymentId: string | undefined) => void;
+	setSelectedSession?: (sessionId: string | undefined) => void;
+	setSessionsSection?: (sessions: SessionSectionViewModel | undefined) => void;
 }