+++ vscode-react/src/interfaces/index.ts
@@ -1,4 +1,1 @@
 export type { IIncomingMessagesHandler } from "@react-interfaces/incomingMessagesHandler.interface";
-// eslint-disable-next-line max-len
-export type { IIncomingDeploymentsMessagesHandler } from "@react-interfaces/incomingDeploymentsMessagesHandler.interface";
-export type { IIncomingSessionsMessagesHandler } from "@react-interfaces/incomingSessionsMessagesHandler.interface";